- entity 변경

+++ daema-api/src/main/java/com/daema/rest/base/dto/CodeDetailDto.java
@@ -10,7 +10,7 @@ import lombok.*;
 @Builder
 public class CodeDetailDto {
 
-	private int codeSeq;
+	private Long codeSeq;
 	private String codeNm;
 	private String codeDesc;
 	private int orderNum;
