+++ daema-api/src/main/java/com/daema/rest/wms/dto/request/StockMoveInsertReqDto.java
@@ -20,7 +20,7 @@ public class StockMoveInsertReqDto {
     private String barcode;
 
     @ApiModelProperty(value = "택배사", example = "0",required = true)
-    private Integer courier;
+    private Long courier;
 
     @ApiModelProperty(value = "송장번호", required = true)
     private String invoiceNo;
