+++ daema-core/src/main/java/com/daema/base/repository/CodeDetailRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CodeDetailRepository extends JpaRepository<CodeDetail, Integer> {
+public interface CodeDetailRepository extends JpaRepository<CodeDetail, Long> {
     List<CodeDetail> findAllByCodeIdInAndUseYnOrderByCodeIdAscOrderNumAsc(List<String> codeIds, String useYn);
 }
