+++ daema-core/src/main/java/com/daema/commgmt/domain/pk/FuncRoleMapPK.java
@@ -12,10 +12,10 @@ public class FuncRoleMapPK implements Serializable {
 	private static final long serialVersionUID = -3997500504328042362L;
 
 	private String funcId;
-	private int roleId;
+	private Long roleId;
 	private Long storeId;
 
-	public FuncRoleMapPK(String funcId, int roleId, Long storeId) {
+	public FuncRoleMapPK(String funcId, Long roleId, Long storeId) {
 		this.funcId = funcId;
 		this.roleId = roleId;
 		this.storeId = storeId;
