+++ daema-core/src/main/java/com/daema/commgmt/repository/RoleMgmtRepository.java
@@ -4,5 +4,5 @@ import com.daema.commgmt.domain.RoleMgmt;
 import com.daema.commgmt.repository.custom.CustomRoleMgmtRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleMgmtRepository extends JpaRepository<RoleMgmt, Integer>, CustomRoleMgmtRepository {
+public interface RoleMgmtRepository extends JpaRepository<RoleMgmt, Long>, CustomRoleMgmtRepository {
 }
