+++ daema-core/src/main/java/com/daema/wms/domain/dto/request/InStockRequestDto.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 public class InStockRequestDto extends SearchParamDto {
     // 통신사
-    private Integer telecom;
+    private Long telecom;
 
     //입고일
     private String inStockRegiDate;
@@ -24,7 +24,7 @@ public class InStockRequestDto extends SearchParamDto {
     private WmsEnum.StockStatStr statusStr;
 
     //제조사
-    private Integer maker;
+    private Long maker;
     
     //기기명
     private Long goodsId;
