+++ daema-core/src/main/java/com/daema/wms/domain/dto/response/InStockResponseDto.java
@@ -39,7 +39,7 @@ public class InStockResponseDto {
      Desc : 기기별 입력정보 및 모델별 입력정보
      */
     // 통신사
-    private int telecom;
+    private Long telecom;
     private String telecomName;
 
     // 공급처
@@ -55,7 +55,7 @@ public class InStockResponseDto {
     private String statusStrMsg;
 
     //제조사
-    private int maker;
+    private Long maker;
     private String makerName;
