+++ daema-core/src/main/java/com/daema/wms/domain/dto/response/InStockWaitDto.java
@@ -42,7 +42,7 @@ public class InStockWaitDto {
      Desc : 기기별 입력정보 및 모델별 입력정보
      */
     // 통신사
-    private int telecom;
+    private Long telecom;
     private String telecomName;
 
     // 공급처
@@ -57,7 +57,7 @@ public class InStockWaitDto {
     private String statusStrMsg;
 
     //제조사
-    private int maker;
+    private Long maker;
     private String makerName;
