+++ daema-core/src/main/java/com/daema/wms/domain/dto/response/ReturnStockResponseDto.java
@@ -44,7 +44,7 @@ public class ReturnStockResponseDto {
      Desc : 기기별 입력정보 및 모델별 입력정보
      */
     // 통신사
-    private int telecom;
+    private Long telecom;
     private String telecomName;
 
     // 공급처
@@ -62,7 +62,7 @@ public class ReturnStockResponseDto {
     private String statusStr;
 
     //제조사
-    private int maker;
+    private Long maker;
     private String makerName;
